+++ cedar-drt/fuzz/fuzz_targets/abac-type-directed-lean.rs
@@ -15,151 +15,12 @@
  */
 
 #![no_main]
+mod abac_type_directed_shared;
+
 use cedar_drt::*;
-use cedar_drt_inner::*;
-use cedar_policy_core::ast;
-use cedar_policy_core::entities::{Entities, NoEntitiesSchema, TCComputation};
-use cedar_policy_core::extensions::Extensions;
-use cedar_policy_generators::{
-    abac::{ABACPolicy, ABACRequest},
-    err::Error,
-    hierarchy::HierarchyGenerator,
-    schema::Schema,
-    settings::ABACSettings,
-};
-use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
-use log::{debug, info};
-use serde::Serialize;
-use std::convert::TryFrom;
+use libfuzzer_sys::fuzz_target;
 
-/// Input expected by this fuzz target:
-/// An ABAC hierarchy, policy, and 8 associated requests
-#[derive(Debug, Clone, Serialize)]
-struct FuzzTargetInput {
-    /// generated schema
-    #[serde(skip)]
-    pub schema: Schema,
-    /// generated entity slice
-    #[serde(skip)]
-    pub entities: Entities,
-    /// generated policy
-    pub policy: ABACPolicy,
-    /// the requests to try for this hierarchy and policy. We try 8 requests per
-    /// policy/hierarchy
-    #[serde(skip)]
-    pub requests: [ABACRequest; 8],
-}
-
-/// settings for this fuzz target
-const SETTINGS: ABACSettings = ABACSettings {
-    match_types: true,
-    enable_extensions: true,
-    max_depth: 3,
-    max_width: 3,
-    enable_additional_attributes: false,
-    enable_like: true,
-    enable_action_groups_and_attrs: true,
-    enable_arbitrary_func_call: true,
-    enable_unknowns: false,
-    enable_action_in_constraints: true,
-    enable_unspecified_apply_spec: true,
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema = Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let policy = schema.arbitrary_policy(&hierarchy, u)?;
-
-        let requests = [
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-            schema.arbitrary_request(&hierarchy, u)?,
-        ];
-        let all_entities = Entities::try_from(hierarchy).map_err(|_| Error::NotEnoughData)?;
-        let entities = drop_some_entities(all_entities, u)?;
-        Ok(Self {
-            schema,
-            entities,
-            policy,
-            requests,
-        })
-    }
-
-    fn size_hint(depth: usize) -> (usize, Option<usize>) {
-        arbitrary::size_hint::and_all(&[
-            Schema::arbitrary_size_hint(depth),
-            HierarchyGenerator::size_hint(depth),
-            Schema::arbitrary_policy_size_hint(&SETTINGS, depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-        ])
-    }
-}
-
-fn drop_some_entities(entities: Entities, u: &mut Unstructured<'_>) -> arbitrary::Result<Entities> {
-    let should_drop: bool = u.arbitrary()?;
-    if should_drop {
-        let mut set: Vec<_> = vec![];
-        for entity in entities.iter() {
-            match u.int_in_range(0..=9)? {
-                0 => (),
-                _ => {
-                    set.push(entity.clone());
-                }
-            }
-        }
-        Ok(Entities::from_entities(
-            set.into_iter(),
-            None::<&NoEntitiesSchema>,
-            TCComputation::AssumeAlreadyComputed,
-            Extensions::all_available(),
-        )
-        .expect("Should be valid"))
-    } else {
-        Ok(entities)
-    }
-}
-
-// The main fuzz target. This is for type-directed fuzzing of ABAC
-// hierarchy/policy/requests
-fuzz_target!(|input: FuzzTargetInput| {
-    initialize_log();
-    let mut policyset = ast::PolicySet::new();
-    policyset.add_static(input.policy.into()).unwrap();
-    let lean_def_engine = LeanDefinitionalEngine::new();
-    debug!("Schema: {}\n", input.schema.schemafile_string());
-    debug!("Policies: {policyset}\n");
-    debug!("Entities: {}\n", input.entities);
-    for request in input.requests.into_iter().map(Into::into) {
-        debug!("Request : {request}");
-        let (rust_res, total_dur) =
-            time_function(|| run_auth_test(&lean_def_engine, request, &policyset, &input.entities));
-
-        info!("{}{}", TOTAL_MSG, total_dur.as_nanos());
-
-        // additional invariant:
-        // type-directed fuzzing should never produce wrong-number-of-arguments errors
-        assert_eq!(
-            rust_res
-                .diagnostics
-                .errors
-                .iter()
-                .map(ToString::to_string)
-                .filter(|err| err.contains("wrong number of arguments"))
-                .collect::<Vec<String>>(),
-            Vec::<String>::new()
-        );
-    }
+fuzz_target!(|input: abac_type_directed_shared::FuzzTargetInput| {
+    let def_engine = LeanDefinitionalEngine::new();
+    abac_type_directed_shared::fuzz(input, &def_engine);
 });
