+++ cedar-drt/fuzz/fuzz_targets/eval-type-directed.rs
@@ -15,138 +15,12 @@
  */
 
 #![no_main]
-use cedar_drt::utils::expr_to_est;
+mod eval_type_directed_shared;
+
 use cedar_drt::*;
-use cedar_drt_inner::*;
-use cedar_policy_core::{
-    ast::Expr,
-    entities::{Entities, NoEntitiesSchema, TCComputation},
-    extensions::Extensions,
-};
-use cedar_policy_generators::abac::ABACRequest;
-use cedar_policy_generators::err::Error;
-use cedar_policy_generators::hierarchy::HierarchyGenerator;
-use cedar_policy_generators::schema::{arbitrary_schematype_with_bounded_depth, Schema};
-use cedar_policy_generators::settings::ABACSettings;
-use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
-use log::debug;
-use serde::Serialize;
-use std::convert::TryFrom;
+use libfuzzer_sys::fuzz_target;
 
-/// Input expected by this fuzz target:
-/// An ABAC hierarchy, policy, and 8 associated requests
-#[derive(Debug, Clone, Serialize)]
-struct FuzzTargetInput {
-    /// generated schema
-    #[serde(skip)]
-    pub schema: Schema,
-    /// generated entity slice
-    #[serde(skip)]
-    pub entities: Entities,
-    /// generated expression
-    #[serde(serialize_with = "expr_to_est")]
-    pub expression: Expr,
-    /// the requests to try for this hierarchy and policy. We try 8 requests per
-    /// policy/hierarchy
-    #[serde(skip)]
-    pub request: ABACRequest,
-}
-
-/// settings for this fuzz target
-const SETTINGS: ABACSettings = ABACSettings {
-    match_types: true,
-    enable_extensions: true,
-    max_depth: 3,
-    max_width: 3,
-    enable_additional_attributes: false,
-    enable_like: true,
-    enable_action_groups_and_attrs: true,
-    enable_arbitrary_func_call: true,
-    enable_unknowns: false,
-    enable_action_in_constraints: true,
-    enable_unspecified_apply_spec: true,
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema = Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let toplevel_type = arbitrary_schematype_with_bounded_depth(
-            &SETTINGS,
-            schema.entity_types(),
-            SETTINGS.max_depth,
-            u,
-        )?;
-        let expr_gen = schema.exprgenerator(Some(&hierarchy));
-        let expression =
-            expr_gen.generate_expr_for_schematype(&toplevel_type, SETTINGS.max_depth, u)?;
-
-        let request = schema.arbitrary_request(&hierarchy, u)?;
-        let all_entities = Entities::try_from(hierarchy).map_err(Error::EntitiesError)?;
-        let entities = drop_some_entities(all_entities, u)?;
-        Ok(Self {
-            schema,
-            entities,
-            expression,
-            request,
-        })
-    }
-
-    fn size_hint(depth: usize) -> (usize, Option<usize>) {
-        arbitrary::size_hint::and_all(&[
-            Schema::arbitrary_size_hint(depth),
-            HierarchyGenerator::size_hint(depth),
-            Schema::arbitrary_policy_size_hint(&SETTINGS, depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-            Schema::arbitrary_request_size_hint(depth),
-        ])
-    }
-}
-
-/// Randomly drops some of the entities from the final list of entities, so we can some invalid derefs
-fn drop_some_entities(entities: Entities, u: &mut Unstructured<'_>) -> arbitrary::Result<Entities> {
-    let should_drop: bool = u.arbitrary()?;
-    if should_drop {
-        let mut set: Vec<_> = vec![];
-        for entity in entities.iter() {
-            match u.int_in_range(0..=9)? {
-                0 => (),
-                _ => {
-                    set.push(entity.clone());
-                }
-            }
-        }
-        Ok(Entities::from_entities(
-            set.into_iter(),
-            None::<&NoEntitiesSchema>,
-            TCComputation::AssumeAlreadyComputed,
-            Extensions::all_available(),
-        )
-        .expect("Should be valid"))
-    } else {
-        Ok(entities)
-    }
-}
-
-// The main fuzz target. This is for type-directed fuzzing of expression evaluation
-fuzz_target!(|input: FuzzTargetInput| {
-    initialize_log();
-    let java_def_engine =
-        JavaDefinitionalEngine::new().expect("failed to create definitional engine");
-    debug!("Schema: {}\n", input.schema.schemafile_string());
-    debug!("expr: {}\n", input.expression);
-    debug!("Entities: {}\n", input.entities);
-    run_eval_test(
-        &java_def_engine,
-        input.request.into(),
-        &input.expression,
-        &input.entities,
-        SETTINGS.enable_extensions,
-    )
+fuzz_target!(|input: eval_type_directed_shared::FuzzTargetInput| {
+    let def_engine = JavaDefinitionalEngine::new().expect("failed to create definitional engine");
+    eval_type_directed_shared::fuzz(input, &def_engine);
 });
